fortesting example

+++ TESTING_TOOL/fortest.py
@@ -2,5 +2,5 @@ def square(a: int, b: int) -> int:
     return a * b
 
 
-def not_square(a: int, b: int) -> int:
+def not_square(a: float, b: float) -> float:
     return a + b
